+++ app/views/application/_pickup_posts.html.haml
@@ -1,5 +1,5 @@
 - if pickup_posts.count >= 1
   %h2.subheading.subheading--aside.subheading--press-articles ピックアップ記事
   %ul
-    - pickup_posts.includes(:post).order(:order).each do |pickup_post|
+    - pickup_posts.includes(:post).ordered.each do |pickup_post|
       = render partial: 'posts/summary_with_content', locals: {post: pickup_post.post}
